refactor: change icon name
