+++ src/frontend/src/lib/components/icons/IconCkConvert.svelte
